EPMRPP-108281 || Refactor retention settings to use long values for seconds and update related logic

+++ src/main/java/com/epam/ta/reportportal/core/project/validator/attribute/OrganizationRetentionLimitValidator.java
@@ -69,7 +69,7 @@ public class OrganizationRetentionLimitValidator {
     return Long.parseLong(value);
   }
 
-  private long orMax(int seconds) {
+  private long orMax(long seconds) {
     return seconds == 0 ? Long.MAX_VALUE : seconds;
   }
 }
